fix: embed field wasn't array, joinDMMessage added, status update on analytics, .env status message
