+++ sample.env
@@ -6,3 +6,4 @@ DISCORD_BOT_TOKEN=The token for the Discord bot.
 DISCORD_BOT_CLIENT_ID=The client ID for the Discord bot.
 DISCORD_INVITE_LINK=The invite link for the Discord bot (Optional, defaults to https://bot.auxdible.me)
 BOT_HOMEPAGE=The homepage for the Discord bot (Optional, defaults to https://bot.auxdible.me)
+BOT_STATUS_MESSAGE=The message you want to show up in the discord bot's status, next to the servers.
