+++ src/interaction/events/guild/guildCreate.ts
@@ -7,7 +7,6 @@ export default async function guildCreate(auxdibot: Auxdibot, guild: Guild) {
    if (!guild) return;
    const channel = guild.systemChannel;
    if (!channel) return;
-   if (auxdibot.updateDiscordStatus) await auxdibot.updateDiscordStatus();
    await findOrCreateServer(auxdibot, guild.id);
    await channel.send({ embeds: [auxdibot.embeds.welcome.toJSON()], components: [promoRow.toJSON()] });
 }
