+++ src/interaction/events/guild/guildDelete.ts
@@ -3,7 +3,6 @@ import { Auxdibot } from '@/interfaces/Auxdibot';
 import deleteServer from '@/modules/server/deleteServer';
 
 export default async function guildDelete(auxdibot: Auxdibot, guild: Guild) {
-   if (auxdibot.updateDiscordStatus) await auxdibot.updateDiscordStatus();
    await deleteServer(auxdibot, guild.id);
    return;
 }
